refactor: use new settings menu on mobile too
