+++ packages/shared/src/components/profile/Header.tsx
@@ -15,7 +15,7 @@ import {
   largeNumberFormat,
   ReferralCampaignKey,
 } from '../../lib';
-import { ProfileSettingsMenu } from './ProfileSettingsMenu';
+import { ProfileSettingsMenuMobile } from './ProfileSettingsMenu';
 import { RootPortal } from '../tooltips/Portal';
 import { GoBackButton } from '../post/GoBackHeaderMobile';
 import { useViewSize, ViewSize } from '../../hooks';
@@ -262,7 +262,7 @@ export function Header({
             aria-label="Edit profile"
           />
           <RootPortal>
-            <ProfileSettingsMenu
+            <ProfileSettingsMenuMobile
               isOpen={isMenuOpen}
               onClose={() => setIsMenuOpen(false)}
             />
