+++ packages/shared/src/components/profile/ProfileSettingsMenu.tsx
@@ -1,210 +1,208 @@
 import type { ReactElement } from 'react';
-import React, { useCallback, useMemo } from 'react';
+import React from 'react';
+import { useRouter } from 'next/router';
+import classNames from 'classnames';
 import {
   AddUserIcon,
   BellIcon,
-  CardIcon,
   EditIcon,
-  LockIcon,
   DevCardIcon,
-  ExitIcon,
   EmbedIcon,
   DocsIcon,
-  TerminalIcon,
   FeedbackIcon,
   AppIcon,
-  DevPlusIcon,
   PrivacyIcon,
   MegaphoneIcon,
+  UserIcon,
+  BlockIcon,
+  CoinIcon,
+  CreditCardIcon,
+  HashtagIcon,
+  HotIcon,
+  InviteIcon,
+  MagicIcon,
+  MailIcon,
+  NewTabIcon,
+  PhoneIcon,
+  ReputationLightningIcon,
+  ExitIcon,
 } from '../icons';
 import { NavDrawer } from '../drawers/NavDrawer';
-import { businessWebsiteUrl, docs, feedback } from '../../lib/constants';
-import { useLazyModal } from '../../hooks/useLazyModal';
-import { LazyModal } from '../modals/common/types';
-import { anchorDefaultRel } from '../../lib/strings';
-import type { NavItemProps } from '../drawers/NavDrawerItem';
+import {
+  appsUrl,
+  businessWebsiteUrl,
+  docs,
+  feedback,
+  reputation,
+  settingsUrl,
+  walletUrl,
+} from '../../lib/constants';
+
+import type { ProfileSectionItemProps } from '../ProfileMenu/ProfileSectionItem';
+import { ProfileSection } from '../ProfileMenu/ProfileSection';
 import { LogoutReason } from '../../lib/user';
 import { useAuthContext } from '../../contexts/AuthContext';
-import { usePrompt } from '../../hooks/usePrompt';
-import { ButtonColor } from '../buttons/Button';
-import { usePlusSubscription } from '../../hooks/usePlusSubscription';
-import { LogEvent, TargetId } from '../../lib/log';
-import { GooglePlayIcon } from '../icons/Google/Play';
-import { checkIsBrowser, isIOSNative, UserAgent } from '../../lib/func';
-import { useConditionalFeature } from '../../hooks';
-import {
-  featureOnboardingAndroid,
-  featurePlusCtaCopy,
-} from '../../lib/featureManagement';
-import { SubscriptionProvider } from '../../lib/plus';
-import { postWebKitMessage, WebKitMessageHandlers } from '../../lib/ios';
+import type { WithClassNameProps } from '../utilities';
+import { HorizontalSeparator } from '../utilities';
+import { useFeatureTheme } from '../../hooks/utils/useFeatureTheme';
+import { ProfileMenuHeader } from '../ProfileMenu/ProfileMenuHeader';
+import { ProfileImageSize } from '../ProfilePicture';
 
-const useMenuItems = (): NavItemProps[] => {
-  const { logout, isAndroidApp } = useAuthContext();
-  const { openModal } = useLazyModal();
-  const { showPrompt } = usePrompt();
-  const { isPlus, plusProvider, logSubscriptionEvent, plusHref } =
-    usePlusSubscription();
-  const {
-    value: { full: plusCta },
-  } = useConditionalFeature({
-    feature: featurePlusCtaCopy,
-    shouldEvaluate: !isPlus,
-  });
-  const { value: appExperiment } = useConditionalFeature({
-    feature: featureOnboardingAndroid,
-    shouldEvaluate: checkIsBrowser(UserAgent.Android) && !isAndroidApp,
-  });
+type MenuItems = Record<
+  string,
+  {
+    title: string | null;
+    items: Record<string, ProfileSectionItemProps>;
+  }
+>;
 
-  const onLogout = useCallback(async () => {
-    const shouldLogout = await showPrompt({
-      title: 'Are you sure?',
-      className: { buttons: 'mt-5 flex-col-reverse' },
-      okButton: { title: 'Logout', color: ButtonColor.Ketchup },
-    });
+const defineMenuItems = <T extends MenuItems>(items: T): T => items;
 
-    if (shouldLogout) {
-      logout(LogoutReason.ManualLogout);
-    }
-  }, [logout, showPrompt]);
-
-  return useMemo(() => {
-    const downloadAndroidApp = appExperiment
-      ? {
-          label: 'Download mobile app',
-          icon: <GooglePlayIcon />,
-          href: process.env.NEXT_PUBLIC_ANDROID_APP,
-          target: '_blank',
-          rel: anchorDefaultRel,
-        }
-      : undefined;
-
-    const items: NavItemProps[] = [
-      {
-        label: 'Profile',
-        isHeader: true,
+export const accountPageItems = defineMenuItems({
+  main: {
+    title: null,
+    items: {
+      profile: {
+        title: 'Profile details',
+        icon: UserIcon,
+        href: `${settingsUrl}/profile`,
       },
-      { label: 'Edit profile', icon: <EditIcon />, href: '/account/profile' },
-    ];
-
-    items.push({
-      label: isPlus ? 'Manage plus' : plusCta,
-      icon: <DevPlusIcon />,
-      href: plusHref,
-      className: isPlus ? undefined : 'text-action-plus-default',
-      target:
-        isPlus && plusProvider === SubscriptionProvider.Paddle
-          ? '_blank'
-          : undefined,
-      onClick: () => {
-        if (
-          isPlus &&
-          isIOSNative() &&
-          plusProvider === SubscriptionProvider.AppleStoreKit
-        ) {
-          postWebKitMessage(WebKitMessageHandlers.IAPSubscriptionManage, null);
-        }
-
-        logSubscriptionEvent({
-          event_name: isPlus
-            ? LogEvent.ManageSubscription
-            : LogEvent.UpgradeSubscription,
-          target_id: TargetId.ProfileDropdown,
-        });
+      account: {
+        title: 'Account',
+        icon: MailIcon,
+        href: `${settingsUrl}/security`,
       },
-    });
-
-    return [
-      ...items,
-      {
-        label: 'Invite friends',
-        icon: <AddUserIcon />,
-        href: '/account/invite',
+      appearance: {
+        title: 'Appearance',
+        icon: NewTabIcon,
+        href: `${settingsUrl}/appearance`,
       },
-      { label: 'Devcard', icon: <DevCardIcon />, href: '/devcard' },
-      { label: 'Privacy', icon: <PrivacyIcon />, href: '/account/privacy' },
-      {
-        label: 'Logout',
-        icon: <ExitIcon />,
-        onClick: onLogout,
+      notifications: {
+        title: 'Notifications',
+        icon: BellIcon,
+        href: `${settingsUrl}/notifications`,
       },
-      {
-        label: 'Manage',
-        isHeader: true,
+      invite: {
+        title: 'Invite Friends',
+        icon: InviteIcon,
+        href: `${settingsUrl}/invite`,
       },
-      {
-        label: 'Appearance',
-        icon: <CardIcon />,
-        href: '/account/appearance',
+    },
+  },
+  billing: {
+    title: 'Billing and Monetization',
+    items: {
+      subscription: {
+        title: 'Subscriptions',
+        icon: CreditCardIcon,
+        href: `${settingsUrl}/subscription`,
       },
-      { label: 'Security', icon: <LockIcon />, href: '/account/security' },
-      {
-        label: 'Notifications',
-        icon: <BellIcon />,
-        href: '/account/notifications',
+      coreWallet: {
+        title: 'Core Wallet',
+        icon: CoinIcon,
+        href: walletUrl,
+        external: true,
       },
-      {
-        label: 'Integrations',
-        icon: <AppIcon />,
-        href: '/account/integrations',
+    },
+  },
+  feed: {
+    title: 'Feed Settings',
+    items: {
+      general: {
+        title: 'General',
+        icon: EditIcon,
+        href: `${settingsUrl}/feed/general`,
       },
-      {
-        label: 'Contribute',
-        isHeader: true,
+      tags: {
+        title: 'Tags',
+        icon: HashtagIcon,
+        href: `${settingsUrl}/feed/tags`,
       },
-      {
-        label: 'Community picks',
-        icon: <DocsIcon />,
-        onClick: () => openModal({ type: LazyModal.SubmitArticle }),
+      sources: {
+        title: 'Content sources',
+        icon: AddUserIcon,
+        href: `${settingsUrl}/feed/sources`,
       },
-      {
-        label: 'Suggest new source',
-        icon: <EmbedIcon />,
-        onClick: () => openModal({ type: LazyModal.NewSource }),
+      preferences: {
+        title: 'Content preferences',
+        icon: AppIcon,
+        href: `${settingsUrl}/feed/preferences`,
       },
-      {
-        label: 'Support',
-        isHeader: true,
+      ai: {
+        title: 'AI superpowers',
+        icon: MagicIcon,
+        href: `${settingsUrl}/feed/ai`,
       },
-      {
-        label: 'Advertise',
+      blocked: {
+        title: 'Blocked content',
+        icon: BlockIcon,
+        href: `${settingsUrl}/feed/blocked`,
+      },
+    },
+  },
+  customization: {
+    title: 'Customization',
+    items: {
+      streaks: {
+        title: 'Streaks',
+        icon: HotIcon,
+        href: `${settingsUrl}/customization/streaks`,
+      },
+      devcard: {
+        title: 'DevCard',
+        icon: DevCardIcon,
+        href: `${settingsUrl}/customization/devcard`,
+      },
+      integrations: {
+        title: 'Integrations',
+        icon: EmbedIcon,
+        href: `${settingsUrl}/customization/integrations`,
+      },
+    },
+  },
+  help: {
+    title: 'Help center',
+    items: {
+      privacy: {
+        title: 'Privacy',
+        icon: PrivacyIcon,
+        href: `${settingsUrl}/privacy`,
+      },
+      reputation: {
+        title: 'Reputation',
+        icon: ReputationLightningIcon,
+        href: reputation,
+        external: true,
+      },
+      advertise: {
+        title: 'Advertise',
+        icon: MegaphoneIcon,
         href: businessWebsiteUrl,
-        icon: <MegaphoneIcon />,
-        target: '_blank',
-        rel: anchorDefaultRel,
+        external: true,
       },
-      downloadAndroidApp,
-      {
-        label: 'Docs',
-        icon: <DocsIcon />,
+      apps: {
+        title: 'Apps',
+        icon: PhoneIcon,
+        href: appsUrl,
+        external: true,
+      },
+      docs: {
+        title: 'Docs',
+        icon: DocsIcon,
         href: docs,
-        target: '_blank',
-        rel: anchorDefaultRel,
+        external: true,
       },
-      {
-        label: 'Changelog',
-        icon: <TerminalIcon />,
-        href: '/sources/daily_updates',
-      },
-      {
-        label: 'Feedback',
-        icon: <FeedbackIcon />,
+      support: {
+        title: 'Support',
+        icon: FeedbackIcon,
         href: feedback,
-        target: '_blank',
-        rel: anchorDefaultRel,
+        external: true,
       },
-    ].filter(Boolean);
-  }, [
-    appExperiment,
-    isPlus,
-    logSubscriptionEvent,
-    onLogout,
-    openModal,
-    plusCta,
-    plusHref,
-    plusProvider,
-  ]);
-};
+    },
+  },
+});
+
+export type AccountPageItemsType = typeof accountPageItems;
 
 interface ProfileSettingsMenuProps {
   isOpen: boolean;
@@ -212,7 +210,42 @@ interface ProfileSettingsMenuProps {
   shouldKeepOpen?: boolean;
 }
 
-export function ProfileSettingsMenu({
+export const InnerProfileSettingsMenu = ({ className }: WithClassNameProps) => {
+  const { asPath } = useRouter();
+  const { logout } = useAuthContext();
+
+  return (
+    <nav className={classNames('flex flex-col gap-2', className)}>
+      {Object.entries(accountPageItems).map(([key, menuItem]) => (
+        <ProfileSection
+          key={key}
+          withSeparator
+          title={menuItem.title}
+          items={Object.entries(menuItem.items).map(
+            ([, item]: [string, ProfileSectionItemProps]) => {
+              return {
+                ...item,
+                isActive: asPath === item.href,
+              };
+            },
+          )}
+        />
+      ))}
+
+      <ProfileSection
+        items={[
+          {
+            title: 'Log out',
+            icon: ExitIcon,
+            onClick: () => logout(LogoutReason.ManualLogout),
+          },
+        ]}
+      />
+    </nav>
+  );
+};
+
+export function ProfileSettingsMenuMobile({
   isOpen,
   onClose,
   shouldKeepOpen,
@@ -225,9 +258,36 @@ export function ProfileSettingsMenu({
         isOpen,
         onClose,
       }}
-      items={useMenuItems()}
-    />
+    >
+      <InnerProfileSettingsMenu className="py-4" />
+    </NavDrawer>
   );
 }
 
-export default ProfileSettingsMenu;
+export function ProfileSettingsMenuDesktop(): ReactElement {
+  const { user } = useAuthContext();
+  const featureTheme = useFeatureTheme();
+
+  if (!user) {
+    return null;
+  }
+
+  return (
+    <aside
+      className={classNames(
+        'ml-auto flex min-h-full flex-col gap-2 self-start rounded-16 border border-border-subtlest-tertiary p-2 tablet:w-64',
+        featureTheme ? 'bg-transparent' : undefined,
+      )}
+    >
+      <ProfileMenuHeader
+        className="rounded-10 px-1 hover:bg-theme-active"
+        shouldOpenProfile
+        profileImageSize={ProfileImageSize.Medium}
+      />
+
+      <HorizontalSeparator />
+
+      <InnerProfileSettingsMenu />
+    </aside>
+  );
+}
